PoC: Add an Akela bridge
This is a proof of concept, that wraps some of the KeyboardioFirmware
things up in a way that make it possible to use some of the auxiliary
Akela libraries, such as Akela-TapDance, with the KeyboardioFirmware.

Because this is just a proof of concept, the implementation is very
barebones, and not very friendly. But it shows this can be done.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>