Fix a warning in led_get_crgb_at
Instead of not handling the case where the LED index is out of range,
and thus producing a warning, return a black color instead. It's as good
as anything else, but at least gets rid of the warning.

And, it goes hand-in-hand with the set counterpart, which, in a similar
case, does nothing: thus the out-of-range LED remains black forever.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>