Split out LED effects into smaller "plugins"
Recreate LEDControl as a class with pluggable parts, similar in vein to
the event handler and loop hooks. Except in this case, only the current
effect runs at any one time, the current one.

All existing effects were separated out into plugins, and the default
firmware example was updated too. All of them were pretty trivial, save
the special NumLock effect: that one also installs a loop hook, and
switches the LED mode if need be. Its setup function also skips to the
next effect, if the mode was selected manually.

Behaviour should be the same as before, but LED effects are now
pluggable, at the cost of some code and data size increase.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>