Implement Model01::led_set_crgb_at(row, col, color)
While this function had a declaration in the header, it lacked an
implementation. With this patch, `key_led_map` moves to the cpp from the
header, and the function gains an implementation.

An additional helper, `led_set_crgb_at(row, col, color)` is added too,
similarly to the other arities.

This makes it a lot easier to address LEDs when knowing the key
position.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>