Add KEYMAP macros for more readable keymap definitions
For those who'd like to work with C arrays, instead of generating
layouts from some other source, the `KEYMAP` macro makes it easier to
define a layer. The parameters in the macro arguments are arranged in
such a way that they can be formatted in a way that matches each hand
more closely. Similarly, a `KEYMAP_STACKED` is introduced, where the
hands are under each other, not side-by-side.

This is just aesthetics, for the C-savvier users.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>