BootAnimation: Only display the animation once
Save the seen state in EEPROM, so that the boot animation is displayed
only on first boot, and not every time the keyboard boots, or is plugged
in.

This way, we still have a nice welcome animation on first boot, but it
won't become an annoying delay every time the keyboard boots.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>