Introduce a helper macro to compute the keymap size
When one adds or removes a layer from a keymap, the keymap size has to
be adjusted in the call to `Keyboardio.setup()`. This is easy to forget,
so as a helper, introduce `KEYMAP_SIZE`, a macro that automatically
computes the size at compile time.

This way, one does not need to remember to update the size anywhere.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>