Implement a way to do simple macros
The main thing here is `Macros_::play`, which takes a list of bytes from
PROGMEM, and plays a macro. The array is always a command, followed by
arguments, and the size of the argument depends on the command: key presses and
releases take a 16-bit argument, and the event is injected into the event
handler flow. Waiting and interval change take a 8-bit time. Helpers are
provided to make it a little bit easier to construct a macro.

Of course, the `macroAction` method may do any other side effects, and is not
restricted to returning a sequence of commands.

Fixes #5.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>