Make modifier macros safer, and use them
Wraps the value LCTRL & friends in parens, so that they can be easily passed on
to other macros without the compiler getting confused.

Also moved the definition up, just below the _HELD bits, so that these macros
can be used for other key definitions, such as `Key_Pipe`, and so on.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>