Rework the hook handling
We can't register hooks from constructors, because there is no guaranteed order
in which the objects will be created. So it may well happen that the Keyboardio
object gets created later, and zeroes out everything. Or it gets created first,
and registers the default handler as the first one, making all the others
pointless.

Instead, we create a KeyboardioPlugin class, that has a `begin` method. This is
responsible for setting up the hooks and whatnot. To make things simpler (for
some values of simple), a `Keyboardio.use` method is introduced, which, when
given a NULL-terminated list of plugin object pointers, will call the begin
method of each.

All LED effects and other plugins that used to register a static object now use
an extern, and had their initialization moved to the `begin` method.

The end result is not the nicest thing, but it works. We can try figuring out
something nicer later.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>