MouseKeys: Make 'em work!
The mousekey handler was way too eager, and captured events that were not meant
to be handled by it, like the `a` key. This has been fixed by removing the
`KEY_MOUSE_CENTER` bit, and replacing it with `KEY_MOUSE_BUTTON`. That way,
everything fits into the `IS_MOUSE_KEY` flag bit.

While there, also fixed the id of the right mouse button.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>