Rearrange the SYNTHETIC bits, and their handling
Introduces the RESERVED bit, a bit reserved for plugins. If it is set, the core
handlers will not handle the event.

Also rearranges the SYNTHETIC bits, to make slightly more sense. In practice,
this means that LED_TOGGLE was promoted to a flag bit, under IS_INTERNAL.

The handler that deals with synthetic events was updated to look at the flag
bits in an order that does not cause confusion, and preparations were also made
to turn it into an independent handler on its own (but that step has not been
taken yet).

This is just groundwork to clean things up, and make the event flow easier to
follow.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>