Loop hooks: Run the loop hooks twice
Some behaviour would benefit from knowing the last report, some would want a
clear report and a way to pre-fill the next one. For this reason, we need two
places to call loop hooks from: one after scanning the matrix and updating LEDs,
and another after sending and clearing the report.

To save space and sanity, we call the same hooks twice, but the second time, we
set the `postClear` flag, which is false the first time.

All users have been updated.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>