Layers: Don't use hooks for the layer event handler
The layer handling is a core functionality, it should be active at all times,
and should be at the very end, before the default event handler. Otherwise there
may be ordering issues, when a plugin wants to return layer keys from its own
event handler.

This also saves us a couple of bytes of both code and data, as an additional
bonus!

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>