Prevent optimising out the hook functions
With the Layer code not using the hooks anymore, the Arduino builder will tell
the linker to remove any unreferenced code. As we are using dot_a_linkage, that
means that the hook functions will be removed due to being unreferenced before
plugins had a chance to reference them.

Add a dummy call in Keyboardio_::setup() to prevent this case.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>