Fix cancellation on interrupt
Cancel the one-shot state when the interruptor is pressed, not when it is
released. Otherwise, when typing fast, other keys pressed while the interruptor
is still releasing, may have the one-shot effect applied too.

For this to work, we need to do the cancellation post-clear, otherwise the
one-shot modifiers are cleared before they register with the interruptor.

Fixes #1, which used to be keyboardio/Akela#121.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>