More reliable modifier tracking
Instead of tracking the modifiers in a loop hook, track them in the event
handler. We catch all modifiers, even injected ones. This makes things more
reliable, because the loop hook can also catch modifiers that TopsyTurvy
injected, which is not desirable.

Oh, yeah, we do not catch TopsyTurvy-injected modifiers, because we use the
Keyboard singleton directly, instead of injecting the keys back into the
workflow.

This reduces the jitter described in #1, but does not fix it on its own.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>