actionKeys: Improved overflow handling
When tapping a tap-dance key handled by `TapDance::actionKeys` more times than
the amount of keys available, treat it as if the last action was to be
triggered. This feels not only more intuitive, but it also allows the counter
and state to reset as soon as possible, resulting in a better typing experience.

Fixes #1.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>