Mask out interruptors
When a key interrupts the one-shot, mask the position out until a key release.
This is needed, because even a short tap usually results in a short hold, which
in turn - without the mask - handles the key.

Instead, we now mask the position out until it is released, so no such problems
arise. Fixes #4.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>