Fix the LED mode handling
First of all, to disallow switching to LEDNumlock, we need to do the check in
the `.init()` method, not in `.setup()`. Second, tracking the previous LED mode
can be done a lot simpler - and a lot reliably.

With these changes, the NumLock LED effect works as it is expected to.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>