Introducing the tuning knobs
Instead of counting loops for the purpose of calculating acceleration, use
timers and steps instead. This means that we can now tune how often the mouse
moves (`speedDelay`), how much it moves when it does (`speed`), how fast
acceleration is (`accelSpeed`), and how often we accelerate (`accelDelay`).

By default, the movement speed is one, and there is no delay, while acceleration
has an 50ms delay, and a speed of one.

But all of these can be tuned at run-time: we can turn off acceleration
completely, or slow down the mouse considerably - the possibilities are almost
endless!

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>