Only sync the LEDs, if there is a change
To further improve the LED performance, sync only when there is a change. We do
this by tracking when change happens, assuming everyone uses the provided
accessors.

While we do a bit of extra work each cycle to do the tracking, that pales in
comparison to what we gain by not having to transfer data needlessly.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>