Implement gamma correction for the LEDs
Instead of using the supplied linear color component values as-is,
convert them to gamma-corrected, non-linear values first. This way, we
can use numbers like 127 to mean half brightness, and have it
automatically translated to the correct, gamma-corrected value.

Table copied from:
 https://learn.adafruit.com/led-tricks-gamma-correction/the-quick-fix

This adds a bit over 256 bytes of code, but makes working with colors a
lot easier.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>