Add a way to display symbols in different colors
While it is possible to play with the ->color property, sometimes one
just wants to override the color once. In this case, saving the previous
value, and changing it back would be an overkill.

Instead, add a few functions that take a color argument as well, and
make the color-less arities use the global property.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>