Turn hooks into linked lists
Instead of pre-allocating 64 items for event-handler and loop
hooks (128 bytes total!), use a linked list instead. This is a little
more memory used per hook, but the memory use scales with the number of
hooks we have. Up until 32 hooks, the linked list wins. At this time, we
do not have that many hooks, not even if all the plugins were in use at
the same time.

Another benefit is that we can have more than 64 hooks now, if so we
wish, and memory is the only limit.

Furthermore, turning both event-handler and loop hooks into a linked
list allowed some code to be cleaned up, so they're only implemented
once (using some careful templating).

Fixes #118.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>