Fix compile errors & drop the END state
We can't ++ enums in C++, it seems, so use a switch statement instead. Also
qualify the EEPROMKeymap object, so we call the object method, instead of trying
to call a method on a class as if it was an object. Oops.

Within this process, drop the END state, it is not required anymore.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>