Move the WDT calls to the hardware plugins
WDT is AVR specific, so it has a much better place in the hardware plugins. Move
it there, and call `KeyboardHardware.setup()` earlier, so it can call
`wdt_disable()` before all the other things it needs.

The delay after WDT disabling moves to the hardware plugin too.

Thanks to @wez and @obra for figuring out what to move where (see
keyboardio/Kaleidoscope#129).

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>