Do gamma correction elsewhere
Instead of doing gamma correction when setting a LED color, do it elsewhere,
when talking to the LED hardware. This way, both the programmer, and the
end-user will deal with the unaltered, raw RGB values, and neither has to care
about when and how gamma correction is applied.

The counterpart of this will be in KeyboardioScanner, that re-introduces gamma
correction on that end.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>