Add an USE_PLUGIN() macro
Does the same thing as `Kaleidoscope.use` (and `USE_PLUGINS` as a consequence),
but for only one plugin. While it is a bit more of a burden to call `USE_PLUGIN`
for all plugins, than it is to list them in `USE_PLUGINS`.

The advantage is that the result is noticeably smaller code. Having both allows
the end-user to choose convenience or small size, depending on their needs.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>