Instead of repeating the highest index, repeat the last one
To mimic the normal repeat, track the last topsy key that was pressed, and if
there are multiple ones active, ignore anything but the youngest. This makes it
possible to hold multiple topsy keys, and have the last one repeat.

It does not handle the topsy+normal chording case, though.

But this fixes #4 in a better way.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>