Simplify the API
Instead of requiring the NumLock key row and column, and the numpad layer index
to be passed to `NumLock.toggle` on every call, derive the first two from
Macros.row and Macros.col respectively, and the latter from a new class
variable, which should be set in the `setup()` method of the sketch.

This way, `NumLock.toggle()` becomes argument-less.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>